Split EntityWithSources into its own file
Problem before split:
source.py imports entity.py which imports source.py

After split:
source.py imports entity.py, entity_with_sources.py imports source.py
